# backend/src/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AssetFlow ERP Backend REST API",
    "version": "1.0.0",
    "description": "Comprehensive REST API documenting all asset registry, allocation, resource booking, maintenance, auditing, and organizational setup endpoints."
  },
  "servers": [
    {
      "url": "https://odooassetflow.onrender.com",
      "description": "Production Render Server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your Supabase JWT access token to authorize request headers."
      }
    },
    "schemas": {
...
      "Allocation": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "assetId": { "type": "string", "format": "uuid" },
          "allocatedToEmployeeId": { "type": "string", "format": "uuid", "nullable": true },
          "allocatedToDepartmentId": { "type": "string", "format": "uuid", "nullable": true },
          "allocatedBy": { "type": "string", "format": "uuid" },
          "allocatedAt": { "type": "string", "format": "date-time" },
          "returnedAt": { "type": "string", "format": "date-time", "nullable": true },
          "status": { "type": "string", "enum": ["active", "returned"] }
        }
      },
      "Booking": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "assetId": { "type": "string", "format": "uuid" },
          "bookedByEmployeeId": { "type": "string", "format": "uuid" },
          "startTime": { "type": "string", "format": "date-time" },
          "endTime": { "type": "string", "format": "date-time" },
          "status": { "type": "string", "enum": ["confirmed", "cancelled", "completed"] },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/departments": {
      "get": {
        "summary": "List Departments",
        "description": "Retrieve all departments.",
        "responses": {
          "200": {
            "description": "List of departments.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string", "format": "uuid" },
                      "name": { "type": "string" },
                      "createdAt": { "type": "string", "format": "date-time" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Department",
        "description": "Restricted to admin / super_admin.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Department created." }
        }
      }
    },
    "/api/employees": {
      "get": {
        "summary": "List Employees",
        "description": "Returns list of employee profiles.",
        "responses": {
          "200": { "description": "List of employees." }
        }
      },
      "post": {
        "summary": "Register Employee",
        "description": "Create a new employee profile. Restricted to admin / super_admin.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["userId", "departmentId", "fullName", "employeeCode"],
                "properties": {
                  "userId": { "type": "string", "format": "uuid" },
                  "departmentId": { "type": "string", "format": "uuid" },
                  "fullName": { "type": "string" },
                  "employeeCode": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Employee created." }
        }
      }
    },
...
    "/api/allocations": {
      "get": {
        "summary": "List Allocations",
        "responses": {
          "200": {
            "description": "List of allocations.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Allocation" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Allocation",
        "description": "Restricted to admin / manager / super_admin.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["assetId"],
                "properties": {
                  "assetId": { "type": "string", "format": "uuid" },
                  "allocatedToEmployeeId": { "type": "string", "format": "uuid" },
                  "allocatedToDepartmentId": { "type": "string", "format": "uuid" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Allocation created." }
        }
      }
    },
...
    "/api/bookings": {
      "get": {
        "summary": "List Resource Bookings",
        "responses": {
          "200": {
            "description": "List of bookings.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Booking" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Resource Booking",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["assetId", "startTime", "endTime"],
                "properties": {
                  "assetId": { "type": "string", "format": "uuid" },
                  "startTime": { "type": "string", "format": "date-time" },
                  "endTime": { "type": "string", "format": "date-time" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Booking slot reserved successfully." },
          "409": { "description": "Booking slot overlap conflict." }
        }
      }
    },
...
    "/api/maintenance": {
      "get": {
        "summary": "List Maintenance Requests",
        "responses": {
          "200": { "description": "List of maintenance requests." }
        }
      },
      "post": {
        "summary": "Create Maintenance Request",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["assetId"],
                "properties": {
                  "assetId": { "type": "string", "format": "uuid" },
                  "notes": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Maintenance request raised." }
        }
      }
    },
...
  }
}